fix (future/past): temporary fix to make it run outside Chrome, too
Initial implementation works fine on Chrome but fails on other browsers. The reason are, of course, date parsing issues. This temporary fix will be re-evaluated later. It may make sense to add a config option to both validators for the date format instead of relying on some magic and fallbacks. Fixes #3.